[REFACT] Removed redundant cmake defs. Updated autobuild. Small changes in code

+++ parser/pe/PEFile.cpp
@@ -106,7 +106,7 @@ void  PEFile::wrap(AbstractByteBuffer *v_buf)
     dataDirEntries[pe::DIR_EXCEPTION] = new ExceptionDirWrapper(this);
     dataDirEntries[pe::DIR_RESOURCE] = new ResourceDirWrapper(this, album);
 
-    for (size_t i = 0; i < pe::DIR_ENTRIES_COUNT; i++) {
+    for (int i = 0; i < pe::DIR_ENTRIES_COUNT; i++) {
         this->wrappers[WR_DIR_ENTRY + i] = dataDirEntries[i];
     }
